+++ src/store/fs/util.rs
@@ -2,7 +2,7 @@ use std::future::Future;
 
 use tokio::{select, sync::mpsc};
 pub(crate) mod entity_manager;
-pub(crate) mod watcher;
+pub(crate) mod watch;
 
 /// A wrapper for a tokio mpsc receiver that allows peeking at the next message.
 #[derive(Debug)]
